+++ src/WaveEq2D.h
@@ -11,22 +11,22 @@
 #include "DiffGrid2D.h"
 #include "Source.h"
 
-using Eigen::MatrixXd;
+using Eigen::MatrixXf;
 
 
 class WaveEq2D : public Model2D
 {
     bool Record;
-    MatrixXd LU, C, C2, U_init;
+    MatrixXf LU, C, C2, U_init;
     DiffGrid2D U0, U1, U2;
     PointSource2D source;
-    // std::vector<MatrixXd> SolData;
+    // std::vector<MatrixXf> SolData;
 
 public:
     WaveEq2D();
     ~WaveEq2D() {}
-    void SetParameters(const MatrixXd& set_C);
-    void SetInit(const MatrixXd& set_U_init);
+    void SetParameters(const MatrixXf& set_C);
+    void SetInit(const MatrixXf& set_U_init);
     void SetSource(const PointSource2D& set_source);
     void SetRecord(bool r);
     void EvalLU(DiffGrid2D& U);
